refactor: quality of life refactor
