+++ src/eventbridge/eventpattern/constants.ts
@@ -1,0 +1,3 @@
+export const OPERATIONS = { STARTS_WITH: "startsWith", INCLUDES: "includes" };
+export const INCLUDES_SEARCH_ELEMENT = "searchElement";
+export const STARTS_WITH_SEARCH_STRING = "searchString";
