+++ src/eventbridge/eventpattern/utils.ts
@@ -1,0 +1,16 @@
+import { BinaryOp } from "../../expression";
+
+export const invertBinaryOperator = (op: BinaryOp): BinaryOp => {
+  switch (op) {
+    case "<":
+      return ">";
+    case "<=":
+      return ">=";
+    case ">":
+      return "<";
+    case ">=":
+      return "<=";
+    default:
+      return op;
+  }
+};
